requirements.txt
beartype==0.18.5
einops==0.8.1
hydra-core==1.3.2
ipython==8.12.3
librosa==0.10.2.post1
mamba_ssm==2.2.2
musdb==0.4.0
museval==0.4.0
numpy==1.24.4
omegaconf==2.3.0
packaging==25.0
pytorch_lightning==2.4.0
rotary_embedding_torch==0.6.4
soundfile==0.12.1
torch==2.4.1
torch_ema==0.3
torchaudio==2.4.1+cu121
tqdm==4.67.1
